Change HTTPRequest to always consume request body.
Some clients and middleware expect that the body is always consumed.
Previous to this change, raising a PublishError exception (for
example) could result in a response being sent without the body
being consumed.  Requiring the application code to ensure that the
body is always consumed seems burdensome.  Using a temporary file
or StringIO object seems simplest and should not significantly
affect performance.