Use Form.is_submitted() to check if Widget should parse.
Pass a reference to the form object to widgets.  Use
Form.is_submitted() to check if form has been submitted.  This
allows query strings to be provided to forms with a POST method
while preventing them from erroneously parsing themselves.
Previously the query string would cause them to think the form
was submitted.