Improve the formatting of Quixote generated pages.
Introduce the errors.format_page() function and use it to generate
all HTML pages produced by Quixote.  It provides a single location
to "monkey-patch" if applications want to customize the look of the
pages.  Improve the wording of the messages on some of the error
pages.  Provide some basic CSS rules to make the pages look a little
more modern.  Use the HTML 5 doctype.