Add the BaseSessionManager class.
This class is useful for applications that wish to implement session
management using some DB system that does not provide a dictionary
like interface.  The SessionManager class contains a lot of mapping
related methods that are not actually necessary to implement in order
to having a working session manager.  BaseSessionManager documents the
essential interface and serves as a more useful base class.